0145 feito

+++ source/c#/lab/lab.Domain/Entities/Bloco/b0/Reg0145EFDC.cs
@@ -5,9 +5,9 @@
         public int Id { get; set; }
         public string Reg { get; set; }
         public int CodIncTrib { get; set; }
-        public int VlRecTot { get; set; }
-        public int VlRecAtiv { get; set; }
-        public int VlRecDemaisAtiv { get; set; }
+        public decimal VlRecTot { get; set; }
+        public decimal VlRecAtiv { get; set; }
+        public decimal VlRecDemaisAtiv { get; set; }
         public string InfoCompl { get; set; }
     }
 }
